Created a user input function
Apparently I have no clue how to work pr's so idk what I did the first time. Also, I have no clue wtf is going on with the remaining code we probably need to meet and review what is going on bc comments do not help. It is too late for this shit.